Testing API whit images
I had to install the Pillow library that is used to handle images inside python projects.
command
pip -m install Pyllow
As we added the image item to the models file, I had to do a migration and confirm it with the commands:
python manage.py makemigrations
python manage.py migrate
Then we just run the server and run the tests.
python manage.py runserver